Allow ID to have Empty values
The only Empty value for an array is one that is zero length.
Effectively this means that we can never have empty IDs that are omitted
by json serialization while they are arrays.

So, switch them to slices. This will break the API, as we can no longer
use `==` to compare two IDs; we'll have to use the new Equals method.