idtype: add TypeFromString.
Sometimes, you have the string representation of a type and you want to
make the actual type out of it. This implements such behaviour.

I chose not to add a k/v map with the string as key to save the memory
footprint as the amount of times you would use this would not justify
adding the extra map over the extra time to access the map values.