Events ownership (#49)
Following @jelmersnoeck suggestions about OwnerData and use a pointer for Data.

I also added Email to Actor and Scope. It can be either a user or a team, maybe a team will have an email contact in the future, but the cli filters empty values if that isn't the case.