New ID type to hold external and internal IDs
This adds a new ID type we can use to reference both internal and external entities.
It sets some constraints on what is allowed in those IDs, but they are pretty flexible.
We will first use this on resources, to replace the User_ID and Team_ID fields with a single Owner_ID.
This will allow us to set external owners on resources in addition to the internal ones.
We also have plans to use this in Billing for subscription ownership.
In Marketplace we will use the Domain portion of this ID to reference partners, and constrain it's usage around them.

Relates manifoldco/engineering#6573