Made some changes to the ID before we really start using it
I didn't like the double \\ on JSON serialization, so this cleans it up a bit with a single / slash
James had a similar request, this makes it more like a url in a path, which could be cool.
Also made IsEmpty() part of the interface, because looking forward I'm realizing I'm going to need it.