Quick fix to have the categories table work with idtypes
See the code for why this was necessary, the idtype's `Collection()` method is used to find the database table and its ability to find the table's name is very limited. Since the `Categories` table was our first table to not have a simple "add a s" plural name, we never noticed the issue.